Inline logistic objective parameter
